refactor: rename function for clarity and improve validation logic
