+++ providers/attio/subscribe.go
@@ -30,7 +30,8 @@ func (c *Connector) EmptySubscriptionResult() *common.SubscriptionResult {
 // PATTERN 2 - Standard/Custom Objects (people, companies, deals, etc.):
 //   - Subscribes using generic "record.*" events (record.created, record.updated, record.deleted)
 //   - Requires object_id filter to target the specific object type
-//   - Object metadata is fetched dynamically from Attio API (Objects can be activated/deactivated in the workspace settings)
+//   - Object metadata is fetched dynamically from Attio API
+//   - Objects can be activated/deactivated in the workspace settings
 //
 // nolint: funlen, cyclop, godoclint
 func (c *Connector) Subscribe(
@@ -43,25 +44,25 @@ func (c *Connector) Subscribe(
 	}
 
 	// Fetch the current list of objects from Attio API
-	objectList, err := c.geStandardOrCustomObjectsList(ctx)
+	objectList, err := c.readStandardOrCustomObjectsList(ctx)
 	if err != nil {
 		return nil, err
 	}
 
 	// Build a map: object name -> object ID
 	// Example: "people" -> "0e80364d-70b1-44d3-b7ba-0a6a564a7152"
-	objectIDMap := make(map[common.ObjectName]string)
+	standardObjects := make(map[common.ObjectName]string)
 	for _, obj := range objectList {
-		objectIDMap[common.ObjectName(obj.ApiSlug)] = obj.Id.ObjectId
+		standardObjects[common.ObjectName(obj.ApiSlug)] = obj.Id.ObjectId
 	}
 
 	// Validate that requested events are supported
-	err = validateSubscriptionEvents(params.SubscriptionEvents, objectIDMap)
+	err = validateSubscriptionEvents(params.SubscriptionEvents, standardObjects)
 	if err != nil {
 		return nil, err
 	}
 
-	payload, err := buildPayload(params.SubscriptionEvents, objectIDMap, req.WebhookEndpoint)
+	payload, err := buildPayload(params.SubscriptionEvents, standardObjects, req.WebhookEndpoint)
 	if err != nil {
 		return nil, fmt.Errorf("failed to build subscription payload: %w", err)
 	}
@@ -164,54 +165,31 @@ func (c *Connector) deleteSubscription(ctx context.Context, subscriptionID strin
 
 func buildPayload(
 	subscriptionEvents map[common.ObjectName]common.ObjectEvents,
-	objectIDMap map[common.ObjectName]string,
+	standardObjects map[common.ObjectName]string,
 	webhookURL string,
 ) (*subscriptionPayload, error) {
 	subscriptions := make([]subscription, 0)
 
 	for objectName, events := range subscriptionEvents {
 		for _, event := range events.Events {
-			EventsMap, isCoreObject := getObjectEvents(objectName)
+			objectEvents, isCoreObject := getObjectEvents(objectName)
+
 			if isCoreObject {
-				providerEvents := EventsMap.toProviderEvents(event)
-
-				if len(providerEvents) == 0 {
-					return nil, fmt.Errorf("%w: no provider events for object '%s' and event '%s'",
-						errUnsupportedSubscriptionEvent, objectName, event)
-				}
-
-				for _, e := range providerEvents {
-					subscriptions = append(subscriptions, subscription{
-						EventType: e,
-						Filter:    nil,
-					})
-				}
-			} else {
-				objectId, exists := objectIDMap[objectName]
-				if !exists {
-					return nil, fmt.Errorf("object '%s' not supported or not activated in workspace", objectName)
-				}
-
-				providerEvent, err := toRecordEvents(event)
+				// Handle building subscriptions for core objects
+				subs, err := buildSubscriptionPayloadForCoreObj(objectEvents, objectName, event)
 				if err != nil {
 					return nil, err
 				}
 
-				filter := map[string]any{
-					"$and": []map[string]any{
-						{
-							"field":    "id.object_id",
-							"operator": "equals",
-							"value":    objectId,
-						},
-					},
+				subscriptions = append(subscriptions, subs...)
+			} else {
+				// Handle building subscriptions for standard/custom objects
+				subs, err := buildSubscriptionPayloadForStandardObj(standardObjects, objectName, event)
+				if err != nil {
+					return nil, err
 				}
-				subscriptions = append(subscriptions,
-					subscription{
-						EventType: providerEvent,
-						Filter:    filter,
-					},
-				)
+
+				subscriptions = append(subscriptions, subs...)
 			}
 		}
 	}
@@ -224,57 +202,4 @@ func buildPayload(
 	}
 
 	return payload, nil
-}
-
-func getObjectEvents(objectName common.ObjectName) (objectEvents, bool) {
-	events, exists := attioObjectEvents[objectName]
-	if !exists {
-		return objectEvents{}, false
-	}
-
-	return events, true
-}
-
-// getAllSupportedEvents returns all provider events that this object supports.
-func (e objectEvents) getAllSupportedEvents() []providerEvent {
-	var events []providerEvent
-
-	events = append(events, e.createEvents...)
-	events = append(events, e.updateEvents...)
-	events = append(events, e.deleteEvents...)
-
-	return events
-}
-
-// getProviderEvents converts a common event type to corresponding Attio provider events.
-func (e objectEvents) toProviderEvents(commonEvent common.SubscriptionEventType) []providerEvent {
-	switch commonEvent { // nolint:exhaustive
-	case common.SubscriptionEventTypeCreate:
-		return e.createEvents
-	case common.SubscriptionEventTypeUpdate:
-		return e.updateEvents
-	case common.SubscriptionEventTypeDelete:
-		return e.deleteEvents
-	default:
-		return nil
-	}
-}
-
-// toRecordEvents converts a common event type to generic Attio record events.
-// This is used for standard/custom objects (people, companies, deals, etc.) which use
-// generic "record.*" event types instead of object-specific events.
-func toRecordEvents(commonEvent common.SubscriptionEventType) (providerEvent, error) {
-	switch commonEvent {
-	case common.SubscriptionEventTypeCreate:
-		return "record.created", nil
-
-	case common.SubscriptionEventTypeUpdate:
-		return "record.updated", nil
-
-	case common.SubscriptionEventTypeDelete:
-		return "record.updated", nil
-
-	default:
-		return "", fmt.Errorf("unsupported event type %s", commonEvent)
-	}
 }
